toque algunas cosas xd pero nada que use alguien mas

+++ CodigoTPO/src/adapters/AdapterTCredito.java
@@ -6,7 +6,7 @@ public class AdapterTCredito implements AdapterMedioDePagoTCredito {
     private float limiteTarjeta;
 
     @Override
-    public void abonarReserva(int nroTar, float total) {
+    public void abonarReserva(long nroTar, float total) {
         if (limiteTarjeta - total >= 0) {
             limiteTarjeta -= total;
             System.out.println("Limite de la tarjeta restante actualizado: " + limiteTarjeta);
