+++ CodigoTPO/src/adapters/interfaces/AdapterMedioDePagoTDebito.java
@@ -2,6 +2,6 @@ package adapters.interfaces;
 
 public interface AdapterMedioDePagoTDebito {
 
-    void abonarReserva(int nroTar, float total);
+    void abonarReserva(long nroTar, float total);
 
 }
